tests/wpt/web-platform-tests/app-history/transition/navigate-transitionWhile-stop.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(() => {
    assert_equals(appHistory.transition, null);
    const from = appHistory.current;

    appHistory.addEventListener("navigate", t.step_func(e => {
      e.transitionWhile(Promise.resolve());
      assert_equals(appHistory.transition, null);
    }));

    let navigateerrorFired = false;
    appHistory.addEventListener("navigateerror", t.step_func(() => {
      navigateerrorFired = true;

      assert_equals(appHistory.transition.navigationType, "push");
      assert_equals(appHistory.transition.from, from);
    }));

    promise_rejects_dom(t, "AbortError", appHistory.navigate("?1").finished).then(t.step_func_done(() => {
      assert_equals(appHistory.transition, null);
      assert_true(navigateerrorFired);
    }));

    assert_equals(appHistory.transition.navigationType, "push");
    assert_equals(appHistory.transition.from, from);

    window.stop();

    assert_equals(appHistory.transition, null);
  }, 0);
}, "transitionWhile()ed navigate() that is stop()ed");
</script>
